typecasted T_IND with short

+++ srcs/vm/funcs/0x02_ld.c
@@ -6,7 +6,7 @@
 /*   By: Gerhard <Gerhard@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/03 13:37:30 by seronen           #+#    #+#             */
-/*   Updated: 2021/02/06 15:01:11 by Gerhard          ###   ########.fr       */
+/*   Updated: 2021/02/06 15:33:24 by Gerhard          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ void    op_ld(t_carriage *carriage)
     if (carrage->stmt->arg_types[0] == 2)
         carriage->regs[reg_slot] = carrage->stmt->args[0];
     else if (carriage->stmt->arg_types[0] === 4)
-        carriage->regs[reg_slot] = convert_4_bytes(&carriage->pos[arg_value_idx])
+        carriage->regs[reg_slot] = convert_4_bytes(&carriage->pos[(short)arg_value_idx])
     if (carrage->stmt->args[0] == 0)
         carriage->carry_flag = 1;
     else
