+++ srcs/vm/funcs/helper_functions/get_arg_value.c
@@ -6,7 +6,7 @@
 /*   By: Gerhard <Gerhard@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/06 14:50:25 by Gerhard           #+#    #+#             */
-/*   Updated: 2021/02/06 15:02:08 by Gerhard          ###   ########.fr       */
+/*   Updated: 2021/02/06 15:35:05 by Gerhard          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ int      get_arg_value(t_carriage *carriage, int arg_value, int i)
     else
     {
         arg_idx_value = carriage->stmt->args[i] % IDX_MOD;
-        arg_value = convert_4_bytes(&carriage->pos[arg_value_idx])
+        arg_value = convert_4_bytes(&carriage->pos[(short)arg_value_idx])
     }
     return (arg_value);
 }
